+++ packages/pangea-sdk/src/main/java/cloud/pangeacyber/pangea/ai_guard/models/ImageAction.java
@@ -1,0 +1,36 @@
+package cloud.pangeacyber.pangea.ai_guard.models;
+
+import com.fasterxml.jackson.annotation.JsonCreator;
+import com.fasterxml.jackson.annotation.JsonValue;
+
+public enum ImageAction {
+	REPORT("report"),
+	BLOCK("block"),
+	UNKNOWN("");
+
+	private final String value;
+
+	ImageAction(final String value) {
+		this.value = value;
+	}
+
+	@JsonValue
+	public String getValue() {
+		return value;
+	}
+
+	@Override
+	public String toString() {
+		return String.valueOf(value);
+	}
+
+	@JsonCreator
+	public static ImageAction fromValue(final String value) {
+		for (var x : ImageAction.values()) {
+			if (x.value.equals(value)) {
+				return x;
+			}
+		}
+		throw new IllegalArgumentException("Unexpected value '" + value + "'");
+	}
+}