+++ packages/pangea-sdk/src/main/java/cloud/pangeacyber/pangea/ai_guard/models/ImageOverride.java
@@ -1,0 +1,32 @@
+package cloud.pangeacyber.pangea.ai_guard.models;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.annotation.JsonInclude.Include;
+import com.fasterxml.jackson.databind.PropertyNamingStrategies;
+import com.fasterxml.jackson.databind.annotation.JsonNaming;
+import java.util.List;
+import lombok.Builder;
+import lombok.Data;
+import lombok.extern.jackson.Jacksonized;
+
+@Builder
+@Data
+@Jacksonized
+@JsonIgnoreProperties(ignoreUnknown = true)
+@JsonInclude(Include.NON_NULL)
+@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
+public class ImageOverride {
+
+	@Builder.Default
+	Boolean disabled = null;
+
+	@Builder.Default
+	ImageAction action = null;
+
+	@Builder.Default
+	List<String> topics = null;
+
+	@Builder.Default
+	Float threshold = null;
+}